Controller logic of switcher relocated to SwitchController. Switching and switching back using this extension.
Hooks adjusted
E2E tests for working with Switcher added.
Fixed login service dependency.
Added functional test.

Altered. When reinitialising the the core functions are used
When logging in from code, a service is on the fly added.

Added service for adding login FE user when in eID
Added service for managing technical logging in.